app/requirements.txt
geoalchemy2
Flask==0.10.1
Flask-Cors==2.1.2
geojson==1.3.1
-e git://github.com/geomet/geomet.git#egg=geomet
psycopg2==2.6.1
requests==2.8.1
Werkzeug==0.11.2
Flask-SQLAlchemy==2.1
gunicorn
